requirements.txt
tuya-iot-py-sdk
requests